+++ scim-schema/src/main/java/org/osiam/resources/scim/MemberRef.java
@@ -23,6 +23,8 @@
 
 package org.osiam.resources.scim;
 
+import java.io.Serializable;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
@@ -30,11 +32,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This class represents a {@link User} or a {@link Group} which are members of an actual {@link Group}
  * 
  * <p>
- * For more detailed information please look at the <a
- * href="http://tools.ietf.org/html/draft-ietf-scim-core-schema-02#section-8">SCIM core schema 2.0, sections 8</a>
+ * For more detailed information please look at the
+ * <a href="http://tools.ietf.org/html/draft-ietf-scim-core-schema-02#section-8">SCIM core schema 2.0, sections 8</a>
  * </p>
  */
-public class MemberRef extends MultiValuedAttribute { // NOSONAR - will be constructed by the builder or jackson
+public class MemberRef extends MultiValuedAttribute implements Serializable { // NOSONAR - will be constructed by the
+                                                                              // builder or jackson
+
+    private static final long serialVersionUID = 2965422671682767189L;
 
     @JsonProperty
     private Type type;
@@ -72,14 +77,15 @@ public class MemberRef extends MultiValuedAttribute { // NOSONAR - will be const
 
     /**
      * Gets the type of the attribute.
-     * 
+     *
      * <p>
-     * For more detailed information please look at the <a href=
-     * "http://tools.ietf.org/html/draft-ietf-scim-core-schema-02#section-3.2" >SCIM core schema 2.0, section 3.2</a>
+     * For more detailed information please look at the
+     * <a href= "http://tools.ietf.org/html/draft-ietf-scim-core-schema-02#section-3.2" >SCIM core schema 2.0, section
+     * 3.2</a>
      * </p>
-     * 
+     *
      * @return
-     * 
+     *
      * @return the actual type
      */
     public Type getType() {
@@ -118,7 +124,8 @@ public class MemberRef extends MultiValuedAttribute { // NOSONAR - will be const
 
     @Override
     public String toString() {
-        return "MemberRef [display=" + getDisplay() +", value=" + getValue() + ", type=" + type + ", primary=" + isPrimary() 
+        return "MemberRef [display=" + getDisplay() + ", value=" + getValue() + ", type=" + type + ", primary="
+                + isPrimary()
                 + ", operation=" + getOperation() + ", ref=" + getReference() + "]";
     }
 
@@ -134,7 +141,7 @@ public class MemberRef extends MultiValuedAttribute { // NOSONAR - will be const
 
         /**
          * builds an Builder based of the given Attribute
-         * 
+         *
          * @param member
          *        existing Attribute
          */
@@ -142,15 +149,15 @@ public class MemberRef extends MultiValuedAttribute { // NOSONAR - will be const
             super(member);
             type = member.type;
         }
-        
+
         /**
          * builds an Builder based of the given {@link User} or {@link Group}
-         * 
+         *
          * @param resource
          *        existing {@link User} or {@link Group}
          */
         public Builder(Resource resource) {
-        	setValue(resource.getId());
+            setValue(resource.getId());
         }
 
         @Override