+++ scim-schema/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Remove unused meta classes and package
 - Deprecate User/Group/Resource#setSchemas(...)
 - Update schema URNs to SCIM draft 19
+- Make User and Group serializable
 
 ## 1.4 - 2015-06-17